pyproject.toml
[project]
name = "contract-intelligence-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "google-cloud-storage>=2.14.0",
    "chromadb>=1.3.5",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.50.1",
    "langchain>=1.0.8",
    "langchain-text-splitters>=1.0.0",
    "markdown>=3.10",
    "openai>=2.8.0",
    "pdfkit>=1.0.0",
    "pydantic[email]>=2.12.4",
    "pymupdf>=1.26.6",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.14.4",
]
